refactor: use override for necessary method
