+++ src/main/java/domain/user/web/UserReadController.java
@@ -18,24 +18,16 @@ public class UserReadController extends AbstractController {
     @Override
     public void doGet(HttpRequest httpRequest, HttpResponse httpResponse) throws IOException {
         User user = UserService.findById(httpRequest.getParameter(User.USER_ID));
-        byte[] body = objectMapper.writeValueAsBytes(user);
-        httpResponse.setHttpStatus(HttpStatus.OK);
-        httpResponse.addHeader(HttpHeader.CONTENT_TYPE, "application/json;charset=utf-8");
-        httpResponse.addHeader(HttpHeader.CONTENT_LENGTH, String.valueOf(body.length));
-        httpResponse.forward(body);
-    }
-
-    @Override
-    public void doPost(HttpRequest httpRequest, HttpResponse httpResponse) throws IOException {
-        httpResponse.setHttpStatus(HttpStatus.METHOD_NOT_ALLOWED);
-        httpResponse.error();
-    }
-
-    @Override
-    public void doPut(HttpRequest httpRequest, HttpResponse httpResponse) {
-    }
-
-    @Override
-    public void doDelete(HttpRequest httpRequest, HttpResponse httpResponse) {
+        if (user == null) {
+            httpResponse.setHttpStatus(HttpStatus.UNAUTHORIZED);
+            httpResponse.error();
+        }
+        if (user != null) {
+            byte[] body = objectMapper.writeValueAsBytes(user);
+            httpResponse.setHttpStatus(HttpStatus.OK);
+            httpResponse.addHeader(HttpHeader.CONTENT_TYPE, "application/json;charset=utf-8");
+            httpResponse.addHeader(HttpHeader.CONTENT_LENGTH, String.valueOf(body.length));
+            httpResponse.forward(body);
+        }
     }
 }
