+++ src/main/java/webserver/HttpStatus.java
@@ -3,6 +3,7 @@ package webserver;
 public enum HttpStatus {
     OK("OK", 200),
     FOUND("Found", 302),
+    UNAUTHORIZED("Unauthorized", 401),
     NOT_FOUND("Not Found", 404),
     METHOD_NOT_ALLOWED("Method Not Allowed", 405),
     INTERNAL_SERVER_ERROR("Internal Server Error", 500);
